+++ Src/Library/changelog.md
@@ -1,3 +1,2 @@
-### CHANGELOG
-- upgrade nswag to latest (.net 7 support)
-- fix swagger response dto property description doesn't work if case not matched.
+### NEW
+- jwt refresh token support